backend/requirements.txt
fastapi[all]
gunicorn
uvicorn
python-dotenv
firebase-admin